+++ elide-spring/elide-spring-boot-autoconfigure/src/main/java/com/yahoo/elide/spring/config/ElideAsyncConfiguration.java
@@ -7,19 +7,22 @@ package com.yahoo.elide.spring.config;
 
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.async.models.AsyncQuery;
+import com.yahoo.elide.async.service.AsyncCleanerService;
 import com.yahoo.elide.async.service.AsyncExecutorService;
 import com.yahoo.elide.async.service.AsyncQueryDAO;
+import com.yahoo.elide.async.service.DefaultAsyncQueryDAO;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * Async Configuration For Elide Services.  Override any of the beans (by defining your own) to change
- * the default behavior.
+ * Async Configuration For Elide Services.  Override any of the beans (by defining your own) 
+ * and setting flags to disable in properties to change the default behavior.
  */
 @Configuration
 @EntityScan(basePackageClasses = AsyncQuery.class)
@@ -39,5 +42,32 @@ public class ElideAsyncConfiguration {
             AsyncQueryDAO asyncQueryDao) {
         return new AsyncExecutorService(elide, settings.getAsync().getThreadPoolSize(),
                 settings.getAsync().getMaxRunTimeMinutes(), asyncQueryDao);
+    }
+
+    /**
+     * Configure the AsyncCleanerService used for cleaning up async query requests.
+     * @param elide elideObject.
+     * @param settings Elide settings.
+     * @return a AsyncCleanerService.
+     */    
+    @Bean
+    @ConditionalOnMissingBean
+    @ConditionalOnProperty(prefix = "elide.async", name = "cleanupEnabled", matchIfMissing = false)
+    public AsyncCleanerService buildAsyncCleanerService(Elide elide, ElideConfigProperties settings,
+            AsyncQueryDAO asyncQueryDao) {
+        return new AsyncCleanerService(elide, settings.getAsync().getMaxRunTimeMinutes(),
+                settings.getAsync().getQueryCleanupDays(), asyncQueryDao);
+    }
+
+    /**
+     * Configure the AsyncQueryDAO used by async query requests.
+     * @param elide elideObject.
+     * @return an AsyncQueryDAO object.
+     */
+    @Bean
+    @ConditionalOnMissingBean
+    @ConditionalOnProperty(prefix = "elide.async", name = "defaultAsyncQueryDAO", matchIfMissing = true)
+    public AsyncQueryDAO buildAsyncQueryDAO(Elide elide) {
+        return new DefaultAsyncQueryDAO(elide, elide.getDataStore());
     }
 }
