--- elide-spring/elide-spring-boot-autoconfigure/src/main/java/com/yahoo/elide/spring/config/ElideAsyncQueryDAOConfiguration.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2019, Yahoo Inc.
- * Licensed under the Apache License, Version 2.0
- * See LICENSE file in project root for terms.
- */
-package com.yahoo.elide.spring.config;
-
-import com.yahoo.elide.Elide;
-import com.yahoo.elide.async.service.AsyncQueryDAO;
-import com.yahoo.elide.async.service.DefaultAsyncQueryDAO;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * Async Configuration For Elide Services.  Override any of the beans (by defining your own) to change
- * the default behavior.
- */
-@Configuration
-@EnableConfigurationProperties(ElideConfigProperties.class)
-@ConditionalOnExpression("${elide.async.defaultAsyncQueryDAO:true}")
-public class ElideAsyncQueryDAOConfiguration {
-
-    /**
-     * Configure the AsyncQueryDAO used by async query requests.
-     * @param elide elideObject.
-     * @return an AsyncQueryDAO object.
-     */
-    @Bean
-    @ConditionalOnMissingBean
-    public AsyncQueryDAO buildAsyncQueryDAO(Elide elide) {
-        return new DefaultAsyncQueryDAO(elide, elide.getDataStore());
-    }
-}
